backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.1

# Redis and caching
redis[hiredis]>=5.0.1
aioredis>=2.0.1

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.10.0

# Background tasks
celery[redis]>=5.3.4

# Blockchain integration
web3>=6.12.0
eth-account>=0.10.0
hedera-sdk-py>=2.50.0

# AI integration
groq>=0.4.0
langchain>=0.0.350

# Data processing
pandas>=2.1.4
numpy>=1.25.2
httpx>=0.25.2

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Development tools
rich>=13.7.0
